core/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
distlib==0.3.7
Django==4.2.7
djangorestframework==3.15.2
dropbox==11.36.0
filelock==3.13.1
pipenv==2023.10.24
platformdirs==3.11.0
python-dotenv==1.0.1
setuptools==68.2.2
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.6
